Create Instrument_Cluster
