--- Test/0804/Instrument_Cluster_v1/instrumentcluster.cpp
@@ -1,15 +1,0 @@
-#include "instrumentcluster.h"
-#include "ui_instrumentcluster.h"
-
-InstrumentCluster::InstrumentCluster(QWidget *parent)
-    : QWidget(parent)
-    , ui(new Ui::InstrumentCluster)
-{
-    ui->setupUi(this);
-}
-
-InstrumentCluster::~InstrumentCluster()
-{
-    delete ui;
-}
-
