+++ Test/0804/instrument_cluster_v1/instrument_cluster_v1.pro
@@ -1,4 +1,4 @@
-QT       += core gui
+QT       += core gui dbus
 
 greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
