+++ Test/0804/instrument_cluster_v1/instrumentcluster.h
@@ -2,6 +2,9 @@
 #define INSTRUMENTCLUSTER_H
 
 #include <QWidget>
+#include <QtDBus/QDBusInterface>
+#include <QDBusReply>
+#include <QTimer>
 
 QT_BEGIN_NAMESPACE
 namespace Ui { class InstrumentCluster; }
@@ -15,7 +18,12 @@ public:
     InstrumentCluster(QWidget *parent = nullptr);
     ~InstrumentCluster();
 
+public Q_SLOTS:
+    void gui_update();
+
 private:
     Ui::InstrumentCluster *ui;
+    QDBusInterface carinfo;
+    QTimer *timer;
 };
 #endif // INSTRUMENTCLUSTER_H
